add-friends-likes: доработал FilmController , FilmService, UserService, UserController, Film и User Storage с учетом замечаний по тестам в Postman

+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -16,9 +16,11 @@ import java.util.*;
 @RequiredArgsConstructor
 @Service
 public class FilmService {
+
     private final FilmStorage filmStorage;
     private final UserStorage userStorage;
     private final LikeStorage likeStorage;
+    private final Set<Long> listLikes = new HashSet<>();
 
     // добавление лайка
     public void addLike(Long filmId, Long userId) {
@@ -26,6 +28,7 @@ public class FilmService {
         if (film != null) {
             if (userStorage.getUserById(userId) != null) {
                 likeStorage.addLike(filmId, userId);
+                film.getLikes().add(userId);
             } else {
                 throw new NotFoundException("Пользователь c ID=" + userId + " не найден!");
             }
@@ -38,7 +41,7 @@ public class FilmService {
     public void removeLike(Long filmId, Long userId) {
         Film film = filmStorage.getFilm(filmId);
         if (film != null) {
-            if (film.getLikes().contains(userId)) {
+            if (film.getLikes() != null && film.getLikes().contains(userId)) {
                 likeStorage.deleteLike(filmId, userId);
             } else {
                 throw new NotFoundException("Лайк от пользователя c ID=" + userId + " не найден!");
@@ -49,7 +52,7 @@ public class FilmService {
     }
 
     // вывод 10 наиболее популярных фильмов по количеству лайков
-    public List<Film> getTopFilms(Integer count) {
+    public List<Film> getPopular(Integer count) {
         if (count < 1) {
             throw new ValidationException("Количество фильмов для вывода не должно быть меньше 1");
         }
